predictions_server/pyproject.toml
[project]
name = "cryptohunter"
version = "0.1.0"
description = ""
authors = [
    { name = "Denis Bobela", email = "youtipie@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "scrapy (>=2.12.0,<3.0.0)",
    "openai (>=1.63.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "scrapy-fake-useragent (>=1.4.4,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "ta (>=0.11.0,<0.12.0)",
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-standardized-errors (>=0.14.1,<0.15.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "uvicorn-worker (>=0.3.0,<0.4.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-celery-beat (>=2.8.0,<3.0.0)",
    "pycryptodome (>=3.22.0,<4.0.0)",
    "torch (>=2.7.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"